enabled multi-use and cache
The way the cache was implemented it only worked for one twitter user.
We needed to use this in a MSM environment for several twitter users,
so I simply added the twitter handle to the cache filename.

Also, when cache_limit was set to 0, the plugin just ignored that and
cached anyway. Part of the issue was that it was comparing the template
tag value to the number 0, and the value is a string "0". The other
issue is that nowhere in the file was it checking the value of
CACHE_LIMIT before it read / write on the cache.